refactor: Url を common モジュールに移した
