+++ build.gradle.kts
@@ -13,7 +13,7 @@ buildscript {
 plugins {
   alias(libs.plugins.androidApplication) apply false
   alias(libs.plugins.androidLibrary) apply false
-  alias(libs.plugins.hilt) apply false
+  alias(libs.plugins.hiltAndroid) apply false
   alias(libs.plugins.jetbrainsKotlinJvm) apply false
   alias(libs.plugins.kotlinAndroid) apply false
   alias(libs.plugins.parcelize) apply false
