+++ core/domain/src/main/java/app/kaito_dogi/mybrary/core/domain/model/MyBook.kt
@@ -1,5 +1,6 @@
 package app.kaito_dogi.mybrary.core.domain.model
 
+import app.kaito_dogi.mybrary.core.common.model.Url
 import kotlinx.serialization.Serializable
 
 @Serializable
